Step 1 - Adding UIText Ammo
I would be more descriptive but I wrote this description 8 times to get
this to push up to github. Basically, I added tracking to when the
player lifted the space key up because allowing them to leave it down to
shot projectiles reduced the ammo much faster than the projectiles being
shot. Now, you have to hit the spacebar each time you want to shoot a
single projectile, but the ammo decrements properly.