Step 4 - Final
Ok! So there have been a lot of small iterations to this project in
which I didn't take into account uploading to github.

I'll start with what I couldn't get to work. Initially, I wanted
continuous damage in 3 ticks on laser but couldn't get it to work so I
made it's bullet have less delay instead and it's longer projectile.
I also wanted to have uiText appear between each level. I know how to do
that know actually but I don't feel that it's that necessary to
implement at this moment and I'm pretty satisfied with the project as
is.

Another thing is that I wanted the bosses individual parts to have their
own health similar to enemy 4 that gibson created, but I didn't really
have enough time to implement that feature and my attempts at copying
his code were fruitless efforts.

[What's different in this version]
- Title Screen with Start and Exit buttons (Please run the build for
complete functionality_

-UIText on beginning of the game

-Ammo is now working correctly when the player obtains more weapons for
the standard blaster

-In the event powerup frequency and during the boss fight not dropping
powerups, a random power up now spawns in the center of the screen every
80 bullets shot. (80 seemed like a good number considering each clip is
100)

-Boss is now available on level 6 (I tried to make him as boss looking
as possible with unity objects, but I'm not very artsy)

-Win text appears on boss being defeated

-After the boss is defeated the game returns to home screen in 5 seconds